# presentation/swagger/swagger.yaml
definitions:
...
  model.Affliction:
    properties:
      alias:
        type: string
      book:
        $ref: '#/definitions/model.Books'
      cost:
        type: number
      cure:
        type: string
      description:
        type: string
      effect:
        type: string
      engName:
        type: string
      frequency:
        type: string
      initialEffect:
        type: string
      mainType:
        $ref: '#/definitions/model.NameAlias'
      name:
        type: string
      onset:
        type: string
      save:
        type: string
      secondaryEffect:
        type: string
      secondaryTypes:
        items:
          $ref: '#/definitions/model.NameAlias'
        type: array
      typeDescription:
        type: string
    type: object
...
  model.Books:
    properties:
      abbreviation:
        type: string
      alias:
        type: string
      name:
        type: string
      order:
        type: integer
    type: object
...
  model.ErrorResponse:
    properties:
      message:
        type: string
      title:
        type: string
    type: object
...
  model.NameAlias:
    properties:
      alias:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: Сайт с информацией для настольной игры Pathfinder
  license:
    name: Чудинов Андрей Дмитриевич
  title: Pathfinder Family
  version: "20220823"
paths:
...
  /ability:
    get:
      description: Характеристики
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Характеристики
      tags:
      - Site
  /ability/{alias}:
    get:
      description: Информация по характеристике
      parameters:
      - description: Alias характеристики
        in: path
        name: alias
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Информация по характеристике
      tags:
      - Site
...
  /actionsInCombat:
    get:
      description: Действия в бою
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Действия в бою
      tags:
      - Site
...
  /afflictions:
    get:
      description: Недуги
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Недуги
      tags:
      - Site
...
  /api/afflictions:
    get:
      description: Список недугов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Affliction'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список недугов
      tags:
      - API
...
  /bigAndLittleCreatures:
    get:
      description: Большие и маленькие существа в бою
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Большие и маленькие существа в бою
      tags:
      - Site
...
  /class/{alias}/bloodline:
    get:
      description: Наследия
      parameters:
      - description: Alias класса
        in: path
        name: alias
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Наследия
      tags:
      - Site
  /class/{alias}/bloodline/{bloodlineAlias}:
    get:
      description: Информация по наследию
      parameters:
      - description: Alias класса
        in: path
        name: alias
        required: true
        type: string
      - description: Alias наследия
        in: path
        name: bloodlineAlias
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Информация по наследию
      tags:
      - Site
...
  /combatModifiers:
    get:
      description: Боевые модификаторы
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Боевые модификаторы
      tags:
      - Site
  /combatStatistics:
    get:
      description: Основные понятия боевой системы
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Основные понятия боевой системы
      tags:
      - Site
...
  /howCombatWorks:
    get:
      description: Как работает система боя
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Как работает система боя
      tags:
      - Site
  /injuryAndDeath:
    get:
      description: Ранения и смерть
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Ранения и смерть
      tags:
      - Site
...
  /movementAndDistance:
    get:
      description: Перемещение, позиция и дистанция
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Перемещение, позиция и дистанция
      tags:
      - Site
...
  /specialAttacks:
    get:
      description: Особые атаки
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Особые атаки
      tags:
      - Site
  /specialInitiativeActions:
    get:
      description: Сдвиг инициативы
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: INTERNAL_ERROR
          schema:
            type: string
      summary: Сдвиг инициативы
      tags:
      - Site
...
swagger: "2.0"